+++ example/spec/syncGranule/SyncGranuleDuplicateHandlingSkipSpec.js
@@ -1,12 +1,10 @@
 const fs = require('fs');
 const path = require('path');
-const { Execution } = require('@cumulus/api/models');
 const {
   aws: {
     headObject,
     parseS3Uri,
-    s3,
-    s3ObjectExists
+    s3
   },
   stringUtils: { globalReplace },
   testUtils: {
@@ -15,20 +13,19 @@ const {
 } = require('@cumulus/common');
 const { api: apiTestUtils, buildAndExecuteWorkflow, LambdaStep } = require('@cumulus/integration-tests');
 const {
+  deleteFolder,
   loadConfig,
   templateFile,
-  uploadTestDataToBucket,
   timestampedTestDataPrefix,
-  deleteFolder
+  uploadTestDataToBucket
 } = require('../helpers/testUtils');
 const {
-  setupTestGranuleForIngest,
-  loadFileWithUpdatedGranuleId
+  loadFileWithUpdatedGranuleId,
+  setupTestGranuleForIngest
 } = require('../helpers/granuleUtils');
-
 const config = loadConfig();
 const lambdaStep = new LambdaStep();
-const workflowName = 'SyncGranule';
+const workflowName = 'SyncGranuleDuplicateSkipTest';
 
 const granuleRegex = '^MOD09GQ\\.A[\\d]{7}\\.[\\w]{6}\\.006\\.[\\d]{13}$';
 const testDataGranuleId = 'MOD09GQ.A2016358.h13v04.006.2016360104606';
@@ -45,17 +42,14 @@ const s3data = [
   '@cumulus/test-data/granules/MOD09GQ.A2016358.h13v04.006.2016360104606_ndvi.jpg'
 ];
 
-describe('When the Sync Granules workflow is configured to overwrite data with duplicate filenames', () => {
-  const testDataFolder = timestampedTestDataPrefix(`${config.stackName}-SyncGranuleSuccess`);
+describe('When the Sync Granule workflow is configured to skip new data when encountering duplicate filenames', () => {
+  const testDataFolder = timestampedTestDataPrefix(`${config.stackName}-IngestGranuleSuccess`);
   const inputPayloadFilename = './spec/syncGranule/SyncGranule.input.payload.json';
   let inputPayload;
   let expectedPayload;
   const collection = { name: 'MOD09GQ', version: '006' };
   const provider = { id: 's3_provider' };
   let workflowExecution;
-
-  process.env.ExecutionsTable = `${config.stackName}-ExecutionsTable`;
-  const executionModel = new Execution();
 
   beforeAll(async () => {
     // upload test data
@@ -94,55 +88,9 @@ describe('When the Sync Granules workflow is configured to overwrite data with d
     expect(workflowExecution.status).toEqual('SUCCEEDED');
   });
 
-  describe('the SyncGranule Lambda function', () => {
-    let lambdaOutput = null;
-    let files;
-    let key1;
-    let key2;
-    const existCheck = [];
-
-    beforeAll(async () => {
-      lambdaOutput = await lambdaStep.getStepOutput(workflowExecution.executionArn, 'SyncGranule');
-      files = lambdaOutput.payload.granules[0].files;
-      key1 = path.join(files[0].fileStagingDir, files[0].name);
-      key2 = path.join(files[1].fileStagingDir, files[1].name);
-      existCheck[0] = await s3ObjectExists({ Bucket: files[0].bucket, Key: key1 });
-      existCheck[1] = await s3ObjectExists({ Bucket: files[1].bucket, Key: key2 });
-    });
-
-    it('receives payload with file objects updated to include file staging location', () => {
-      expect(lambdaOutput.payload).toEqual(expectedPayload);
-    });
-
-    // eslint-disable-next-line max-len
-    it('receives meta.input_granules with files objects updated to include file staging location', () => {
-      expect(lambdaOutput.meta.input_granules).toEqual(expectedPayload.granules);
-    });
-
-    it('receives files with custom staging directory', () => {
-      files.forEach((file) => {
-        expect(file.fileStagingDir).toMatch('custom-staging-dir\/.*');
-      });
-    });
-
-    it('adds files to staging location', () => {
-      existCheck.forEach((check) => {
-        expect(check).toEqual(true);
-      });
-    });
-  });
-
-  describe('the sf-sns-report task has published a sns message and', () => {
-    it('the execution record is added to DynamoDB', async () => {
-      const record = await executionModel.get({ arn: workflowExecution.executionArn });
-      expect(record.status).toEqual('completed');
-    });
-  });
-
   describe('and it encounters data with a duplicated filename', () => {
     let lambdaOutput;
     let existingfiles;
-    let fileUpdated;
 
     beforeAll(async () => {
       lambdaOutput = await lambdaStep.getStepOutput(workflowExecution.executionArn, 'SyncGranule');
@@ -155,7 +103,6 @@ describe('When the Sync Granules workflow is configured to overwrite data with d
       // update one of the input files, so that the file has different checksum
       const content = randomString();
       const file = inputPayload.granules[0].files[0];
-      fileUpdated = file.name;
       const updateParams = {
         Bucket: config.bucket, Key: path.join(file.path, file.name), Body: content
       };
@@ -172,7 +119,7 @@ describe('When the Sync Granules workflow is configured to overwrite data with d
       expect(workflowExecution.status).toEqual('SUCCEEDED');
     });
 
-    it('overwrites the existing file with the new data', async () => {
+    it('does not overwrite existing file or create a copy of new file', async () => {
       lambdaOutput = await lambdaStep.getStepOutput(workflowExecution.executionArn, 'SyncGranule');
       const files = lambdaOutput.payload.granules[0].files;
 
@@ -181,15 +128,8 @@ describe('When the Sync Granules workflow is configured to overwrite data with d
         return { filename: f.filename, fileSize: header.ContentLength, LastModified: header.LastModified };
       }));
 
-      expect(currentFiles.length).toBe(existingfiles.length);
-
-      currentFiles.forEach((cf) => {
-        const existingfile = existingfiles.filter((ef) => ef.filename === cf.filename);
-        expect(cf.LastModified).toBeGreaterThan(existingfile[0].LastModified);
-        if (cf.filename.endsWith(fileUpdated)) {
-          expect(cf.fileSize).toBe(inputPayload.granules[0].files[0].fileSize);
-        }
-      });
+      expect(currentFiles).toEqual(existingfiles);
+      expect(lambdaOutput.payload).toEqual(expectedPayload);
     });
   });
 });